# core-banking/api/src/main/resources/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "Core banking",
    "description" : "Socle core banking : comptes, operations, credit, parametrage, arretes, restitutions et etats financiers. Toute reponse porte l'enveloppe data / page / error / meta ; les montants sont des chaines avec leur devise ; les actes a double validation repondent 202 puis s'approuvent par un second porteur."
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/v1/entities/{legalEntityId}/accounts/{accountId}/payment-orders" : {
      "post" : {
        "operationId" : "Payment.order",
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.PaymentOrderRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Rejeu d'une cle d'idempotence deja traitee : le premier recu, rien n'est comptabilise deux fois",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "201" : {
            "description" : "Operation comptabilisee : le recu",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
...
    "/v1/entities/{legalEntityId}/payment-orders" : {
      "get" : {
        "operationId" : "Payment.list",
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "$ref" : "#/components/parameters/Page"
        }, {
          "$ref" : "#/components/parameters/PageSize"
        } ],
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                      }
                    },
                    "page" : {
                      "$ref" : "#/components/schemas/Page"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/payment-orders/{orderId}" : {
      "get" : {
        "operationId" : "Payment.read",
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/payment-orders/{orderId}/cancellation" : {
      "post" : {
        "operationId" : "Payment.cancel",
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.Reason"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/payment-orders/{orderId}/return" : {
      "post" : {
        "operationId" : "Payment.returnOrder",
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.Reason"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/payment-orders/{orderId}/send" : {
      "post" : {
        "operationId" : "Payment.send",
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
    "/v1/entities/{legalEntityId}/payment-orders/{orderId}/settlement" : {
      "post" : {
        "operationId" : "Payment.settle",
        "tags" : [ "Payment" ],
        "parameters" : [ {
          "$ref" : "#/components/parameters/RequestId"
        }, {
          "name" : "legalEntityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Requests.Settlement"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Succes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "data" : {
                      "$ref" : "#/components/schemas/PaymentService.PaymentOrder"
                    },
                    "page" : {
                      "type" : "null"
                    },
                    "error" : {
                      "type" : "null"
                    },
                    "meta" : {
                      "$ref" : "#/components/schemas/Meta"
                    }
                  },
                  "required" : [ "data", "page", "error", "meta" ]
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/400"
          },
          "401" : {
            "$ref" : "#/components/responses/401"
          },
          "403" : {
            "$ref" : "#/components/responses/403"
          },
          "404" : {
            "$ref" : "#/components/responses/404"
          },
          "409" : {
            "$ref" : "#/components/responses/409"
          },
          "422" : {
            "$ref" : "#/components/responses/422"
          },
          "500" : {
            "$ref" : "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "description" : "Jeton d'acces Keycloak : l'entite, l'agence et les roles de l'appelant en sont derives, jamais d'un parametre.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "type" : "http"
      }
    },
    "parameters" : {
...
      "IdempotencyKey" : {
        "name" : "Idempotency-Key",
        "in" : "header",
        "required" : true,
        "schema" : {
          "type" : "string"
        },
        "description" : "Cle d'idempotence : un rejeu repond 200 avec le premier recu"
      },
      "Page" : {
        "name" : "page",
        "in" : "query",
        "required" : false,
        "schema" : {
          "default" : 0,
          "type" : "integer",
          "minimum" : 0
        },
        "description" : "Numero de page, a partir de zero"
      },
      "PageSize" : {
        "name" : "size",
        "in" : "query",
        "required" : false,
        "schema" : {
          "minimum" : 1,
          "maximum" : 200,
          "default" : 50,
          "type" : "integer"
        },
        "description" : "Taille de page ; au-dela du maximum, 400"
      },
      "RequestId" : {
        "name" : "X-Request-Id",
        "in" : "header",
        "required" : false,
        "schema" : {
          "type" : "string",
          "pattern" : "^[A-Za-z0-9._:-]{1,64}$"
        },
        "description" : "Identifiant de requete du client, repris dans meta.requestId et en en-tete de reponse ; attribue s'il est absent ou mal forme"
      }
    },
    "responses" : {
      "400" : {
        "description" : "Requete invalide : page hors bornes, parametre mal forme, cle d'idempotence absente, corps illisible, curseur invalide",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "401" : {
        "description" : "Aucun jeton valide",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "403" : {
        "description" : "Habilitation refusee, ou jeton insuffisant",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "404" : {
        "description" : "Ressource inconnue — ou d'une autre entite, que le cloisonnement ne montre pas",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "409" : {
        "description" : "Conflit d'etat : compte bloque, disponible insuffisant, traitement refuse, exercice ou maquette dans un autre etat",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "422" : {
        "description" : "Requete que le socle ne peut pas honorer : devise, montant, parametrage incomplet ou invalide",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "500" : {
        "description" : "Erreur non prevue : rien n'a ete comptabilise",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas" : {
...
      "Error" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          }
        },
        "required" : [ "type", "title", "status" ]
      },
      "ErrorEnvelope" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "null"
          },
          "page" : {
            "type" : "null"
          },
          "error" : {
            "$ref" : "#/components/schemas/Error"
          },
          "meta" : {
            "$ref" : "#/components/schemas/Meta"
          }
        },
        "required" : [ "data", "page", "error", "meta" ]
      },
...
      "Meta" : {
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requestId" : {
            "type" : "string"
          }
        },
        "required" : [ "timestamp", "requestId" ]
      },
      "Money" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "string",
            "description" : "Montant en chaine, a l'echelle de la devise ; jamais un flottant"
          },
          "currency" : {
            "type" : "string",
            "description" : "Code ISO 4217"
          }
        },
        "required" : [ "amount", "currency" ]
      },
...
      "Page" : {
        "type" : "object",
        "properties" : {
          "number" : {
            "type" : [ "integer", "null" ]
          },
          "size" : {
            "type" : "integer"
          },
          "totalElements" : {
            "type" : [ "integer", "null" ]
          },
          "totalPages" : {
            "type" : [ "integer", "null" ]
          },
          "hasNext" : {
            "type" : "boolean"
          },
          "hasPrevious" : {
            "type" : "boolean"
          },
          "nextCursor" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "size", "hasNext", "hasPrevious" ]
      },
...
      "PaymentService.PaymentOrder" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "legalEntityId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "accountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "$ref" : "#/components/schemas/Money"
          },
          "fee" : {
            "$ref" : "#/components/schemas/Money"
          },
          "tax" : {
            "$ref" : "#/components/schemas/Money"
          },
          "beneficiaryName" : {
            "type" : "string"
          },
          "beneficiaryBank" : {
            "type" : "string"
          },
          "beneficiaryAccount" : {
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          },
          "channel" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "orderedOn" : {
            "type" : "string",
            "format" : "date"
          },
          "entryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "clearingAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sentOn" : {
            "type" : "string",
            "format" : "date"
          },
          "settledOn" : {
            "type" : "string",
            "format" : "date"
          },
          "settlementAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "settlementEntryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "returnedOn" : {
            "type" : "string",
            "format" : "date"
          },
          "returnEntryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "returnReason" : {
            "type" : "string"
          },
          "cancelledOn" : {
            "type" : "string",
            "format" : "date"
          },
          "cancelEntryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "cancelReason" : {
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
      "Requests.PaymentOrderRequest" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "beneficiaryName" : {
            "type" : "string"
          },
          "beneficiaryBank" : {
            "type" : "string"
          },
          "beneficiaryAccount" : {
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          },
          "channel" : {
            "type" : "string"
          }
        }
      },
...
      "Requests.Reason" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string"
          }
        }
      },
...
      "Requests.Settlement" : {
        "type" : "object",
        "properties" : {
          "nostroAccountId" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
    }
  }
}